Complete Incorporating Changes for Model Interfaces for Testing
Although the changes are tested and added everywhere, the current
design has a serious flaw that it needs conversion of every model
into an interface. This limits field access to each model struct.

The future, and probably the correct way to go would be to abstract
all the sql features using an interface or a library that does it
for us (say, sqlmock). This needs some research in the community
and then we should proceed with the idiomatic approach