Notify users of Log package by printing fallback log line
Previously, if the caller of the log package did not `Initialize`
the package, then all the log functions would silently fail which
was a design flaw.

This has been fixed by introducing a _fallback mechanism_ in the
package. A fallbackMetaLogger is initialized in the package as a
global variable. This global variable is safe to use since it does
not have any data to tamper with but only a behavior -
`basePrinter(v ...interface{})`. This basePrinter prints a static
message notifying the package needs initialization.

Tests have been added for `isInitialized()` and for the scenario
where log functions are called without initiliazation