Complete implementation of ContentType Middlewares
Now, the incoming request is validated amongst a list of valid content
types and if the accept request header has valid value, then the
value is set in the request context using a custom context so that
the request handler can quickly act on it withou parsing the request.
The content type is also set in the response so that handlers do not have
to do it.

If the requested content type is not supported, the response header
is set with text/plain